Move the location of integration test
